+++ R/utils.R
@@ -298,3 +298,7 @@ round.data.table <- function(x, digits = 2L){
 
 
 .list_corpora = function() toupper(list.files( registry() ))
+
+#' @importFrom magrittr %>%
+#' @export
+magrittr::`%>%`